On branch main  - add docker file

+++ src/utils/helper/storage.helper.js
@@ -1,13 +1,16 @@
+/* eslint-disable no-constant-condition */
 const ACCESS_TOKEN = 'access_token';
 
 export const getToken = () => {
-  return JSON.parse(localStorage.getItem(ACCESS_TOKEN));
+  if (typeof window !== 'undefined')
+    return JSON.parse(localStorage.getItem(ACCESS_TOKEN));
 };
 
 export const setToken = (token) => {
-  localStorage.setItem(ACCESS_TOKEN, JSON.stringify(token));
+  if (typeof window !== 'undefined')
+    localStorage.setItem(ACCESS_TOKEN, JSON.stringify(token));
 };
 
 export const clearStorage = () => {
-  localStorage.clear();
+  if (typeof window !== 'undefined') localStorage.clear();
 };
